CRI: Move truncindex to pkg
