+++ pkg/truncindex/truncindex.go
